Refactor script into separate files
